Change buttons and implement 'add' operation on input
Changed 'Undo' and 'Clear all' buttons to '.' and '=' buttons.
Added functionality to perform add operation on user input.
Changed and added global variables. All class names are now global
variables.
Added functionality to change display according to user state of
calculation.

Example:
When adding 1 and 2:
When 1 is selected, display shows 1.
When plus is pressed and 2 is pressed, display is cleared and shows 2.
When equals is pressed, display is cleared and result is shown.

Note: Currently, only works for add operation on single digit integers.